+++ test/cli/test_commands.py
@@ -36,8 +36,9 @@ def test_commands_help(cli):
 
     assert result.exit_code == ExitCode.OK, result.stdout
     lines = result.stdout.split("\n")
-    assert lines[11] == "  replay  Replay requests from a saved cassette."
-    assert lines[12] == "  run     Perform schemathesis test."
+    assert lines[11] == "  auth    Authenticate Schemathesis.io."
+    assert lines[12] == "  replay  Replay requests from a saved cassette."
+    assert lines[13] == "  run     Perform schemathesis test."
 
     result_help = cli.main("--help")
     result_h = cli.main("-h")